ToDo #11 Implementamos metodos para hallar mínimo, máximo y media de gimnasios por manager

+++ src/main/java/services/AdminService.java
@@ -1,9 +1,11 @@
 package services;
 
 import java.util.Collection;
+import java.util.Iterator;
 
 import javax.transaction.Transactional;
 
+import org.apache.commons.lang.NullArgumentException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,6 +22,9 @@ public class AdminService {
 
     @Autowired
     private AdminRepository adminRepository;
+
+    @Autowired
+    private ManagerService managerService;
 
     /**
      * Will create an Admin with no data.
@@ -40,26 +45,37 @@ public class AdminService {
     }
 
     /**
+     * This method will find an admin by his id from database
      *
-     * @param adminId
-     * @return
+     * @param adminId The admin ID from database
+     * @return An admin, or null if it does not find anything.
      */
     public Admin findOne(int adminId) {
 	return this.adminRepository.findOne(adminId);
     }
 
     /**
+     * This method will save an admin inside database if it does not previously
+     * exists.
      *
-     * @param admin
-     * @return
+     * @param admin The admin to be saved
+     * @return The saved admin
      */
-    public Admin save(Admin admin) {
-	return this.adminRepository.save(admin);
+    public Admin save(Admin admin) throws Exception {
+
+	if (admin == null)
+	    throw new NullArgumentException("admin");
+
+	if (!this.adminRepository.findAll().contains(admin))
+	    return this.adminRepository.save(admin);
+	else
+	    throw new Exception("The admin already exists inside database");
     }
 
     /**
+     * This method will delete an admin from the database.
      *
-     * @param admin
+     * @param admin The admin to be deleted
      */
     public void delete(Admin admin) {
 	this.adminRepository.delete(admin);
@@ -68,26 +84,137 @@ public class AdminService {
     /**
      * Will ban a manager
      *
-     * @param manager
+     * @param manager The manager to be banned
      */
     public void banManager(Manager manager) {
 
+	if (manager != null) {
+
+	    manager.setBanned(true);
+
+	    this.managerService.save(manager);
+	}
     }
 
+    /**
+     * Will unban a manager
+     *
+     * @param manager
+     */
     public void unbanManager(Manager manager) {
 
+	if (manager != null) {
+
+	    manager.setBanned(false);
+
+	    this.managerService.save(manager);
+	}
     }
 
+    /**
+     * This method will return the minimum number of gyms that a manager is in
+     * charge of.<br>
+     * <br>
+     *
+     * By querying all managers stored inside the database, this method checks how
+     * many gyms each manager is in charge of, returning the minimum number of gyms
+     * per manager.
+     *
+     * @return The minimum number of gyms per manager
+     */
     public int getMinimumNumberOfGymsPerManager() {
-	return 0;
+
+	Collection<Manager> managers = this.managerService.findAll();
+
+	if (managers.isEmpty())
+	    return 0;
+
+	// Get an iterator to iterate over this collection of managers
+	Iterator<Manager> it = managers.iterator();
+
+	// Because the minimum number of gyms is wanted, we first initialize
+	// our minimum to the maximum possible integer value. Hence, when
+	// the first manager is checked, no matter how low the value is, it
+	// will establish itself as the minimum value.
+	int minimumNumberOfGyms = Integer.MAX_VALUE;
+
+	while (it.hasNext()) {
+
+	    Manager manager = it.next();
+
+	    if (manager.getGyms().size() < minimumNumberOfGyms)
+		minimumNumberOfGyms = manager.getGyms().size();
+	}
+
+	return minimumNumberOfGyms;
     }
 
+    /**
+     * This method will return the maximum number of gyms per manager.<br>
+     * <br>
+     *
+     * By querying all managers stored inside the database, this method check how
+     * many gyms each manager is in charge of, returning the maximum number of gyms
+     * per manager.
+     *
+     * @return The maximum number of gyms per manager.
+     */
     public int getMaximumNumberOfGymsPerManager() {
-	return 0;
+
+	Collection<Manager> managers = this.managerService.findAll();
+
+	// Get an iterator to iterate over this collection of managers
+	Iterator<Manager> it = managers.iterator();
+
+	int maximumNumberOfGyms = 0;
+
+	while (it.hasNext()) {
+
+	    Manager manager = it.next();
+
+	    if (manager.getGyms().size() > maximumNumberOfGyms)
+		maximumNumberOfGyms = manager.getGyms().size();
+	}
+
+	return maximumNumberOfGyms;
     }
 
-    public double getAverageNumberOfGymsPerManager() {
-	return 0;
+    /**
+     * This method will return the average of how many gyms a manager is in charge
+     * of.<br>
+     * <br>
+     *
+     * Take into account that this method will only return the average integer
+     * portion, leaving out all float values.
+     *
+     * @return The average number of gyms per manager.
+     */
+    public int getAverageNumberOfGymsPerManager() {
+
+	Collection<Manager> managers = this.managerService.findAll();
+
+	if (managers.isEmpty())
+	    return 0;
+
+	// Get an iterator to iterate over this collection of managers
+	Iterator<Manager> it = managers.iterator();
+
+	int totalNumberOfGyms = 0;
+
+	int totalNumberOfManagers = 0;
+
+	while (it.hasNext()) {
+
+	    Manager manager = it.next();
+
+	    totalNumberOfGyms += manager.getGyms().size();
+
+	    totalNumberOfManagers++;
+	}
+
+	int averageNumberOfGymsPerManager = totalNumberOfGyms / totalNumberOfManagers;
+
+	return averageNumberOfGymsPerManager;
     }
 
     public double getStandarDeviationOfGymsPerManager() {
